sir_ali_aftab_assignments/ai_travel_agent/pyproject.toml
[project]
name = "ai-travel-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.6.2",
    "openai-agents>=0.2.3",
    "python-dotenv>=1.1.1",
]
